Refactor to offer a pure python API
This is for decoupling functionality from Django views and the Django
request object. It allows running the whole payment process in a script
with the adyen part mocked. This is useful for automated tests.

Also fixes a severe bug where after payment the order would include items
added to the user's basket after the payment process was started. This
would allow users to add items to their order that they didn't pay for.

In the process, the interface to oscar was also adapted. The oscar changes
are in https://github.com/django-oscar/django-oscar/pull/1678