Upgrade to Django 1.9
In Django 1.9 it is not allowed to import models in __init__.py of the
app module, because it needs to be able to import the module before any
model initialization is performed.

As such I moved the url configuration into app.py, which now has to be
imported to get the urls instead of django_adyen or oscar_adyen
respectively.